requirements.txt
torch
torchvision
transformers
numpy
Pillow
tqdm
matplotlib
